Update matrice ennesima.py

+++ matrice ennesima.py
@@ -4,7 +4,7 @@ y = int(input("Inserisci numero colonne"))
 for i in range (x):
      num = []
      for j in range (y) :
-          n = int(input("Inserisci un numero"))
+          n = float(input("Inserisci un numero"))
           num.insert (i,n)
      matrice.append(num)
